requirements.txt
backports-abc==0.5
certifi==2018.4.16
chardet==3.0.4
Django==1.11.13
futures==3.2.0
idna==2.6
nose==1.3.7
pip-save==0.2.0
python-crontab==2.3.0
python-dateutil==2.7.2
requests==2.18.4
singledispatch==3.4.0.3
six==1.9.0
tornado==5.0.2
urllib3==1.22
